requirements.txt
mkl==2024.0
mmcv==1.7.0
diffusers==0.32.1
timm==1.0.12
accelerate==1.2.1
tensorboard==2.17.0
transformers==4.47.1
ftfy==6.3.1
protobuf==3.20.2
gradio==4.1.1
yapf==0.40.1
bs4==0.0.2
einops==0.8.0
optimum==1.23.3
scipy==1.13.1
Pillow==10.4.0
matplotlib==3.9.4
omegaconf
datasets
sentencepiece~=0.1.99
peft==0.14.0
beautifulsoup4==4.12.3
bitsandbytes
wandb
